+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "orx-concurrent-bag"
-version = "2.4.0"
+version = "2.5.0"
 edition = "2021"
 authors = ["orxfun <orx.ugur.arikan@gmail.com>"]
 description = "An efficient, convenient and lightweight grow-only concurrent data structure allowing high performance concurrent collection."
@@ -11,10 +11,10 @@ categories = ["data-structures", "concurrency", "rust-patterns"]
 
 [dependencies]
 orx-pseudo-default = "1.4"
-orx-pinned-vec = "3.4"
-orx-fixed-vec = "3.4"
-orx-split-vec = "3.4"
-orx-pinned-concurrent-col = "2.4"
+orx-pinned-vec = "3.6"
+orx-fixed-vec = "3.6"
+orx-split-vec = "3.6"
+orx-pinned-concurrent-col = "2.5"
 
 [dev-dependencies]
 criterion = "0.5.1"
@@ -22,6 +22,7 @@ rand = "0.8.5"
 rayon = "1.9.0"
 test-case = "3.3.1"
 append-only-vec = "0.1.5"
+boxcar = "0.2.5"
 
 [[bench]]
 name = "collect_with_extend"
